Move relevant files to Durable\Actions folder
