+++ src/Durable/Actions/OrchestrationAction.cs
@@ -5,7 +5,7 @@
 
 #pragma warning disable 1591 // Missing XML comment for publicly visible type or member 'member'
 
-namespace Microsoft.Azure.Functions.PowerShellWorker.Durable
+namespace Microsoft.Azure.Functions.PowerShellWorker.Durable.Actions
 {
     /// <summary>
     /// Base class that represents an orchestration action.
