+++ src/Durable/DurableTimerTask.cs
@@ -9,6 +9,8 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.Durable
 {
     using System;
     using System.Linq;
+    
+    using Microsoft.Azure.Functions.PowerShellWorker.Durable.Actions;
 
     // Returned by the Start-DurableTimer cmdlet if the NoWait flag is present, representing a timeout task
     // All DurableTimerTasks must be complete or canceled for the orchestration to complete
